+++ src/com/palmergames/util/TimeTools.java
@@ -38,7 +38,7 @@ public class TimeTools {
 		if (dhms.contains("s")) {
 			seconds = Integer.parseInt(dhms.split("s")[0].replaceAll(" ", ""));
 		}
-		return (days * 86400) + (hours * 3600) + (minutes * 60) + seconds;
+		return (days * 86400) + (hours * 3600) + (minutes * 60) + (long)seconds;
 	}
 
 	public static long getMillis(String dhms) {
